Refactoring for shared/common/dmi

+++ src/os/shared/uefibase/test/com/ngos/shared/uefibase/sections/section0/com/ngos/shared/common/dmi/lib/entry/lib/dmibiosextendedromsize.h
@@ -1,74 +1,74 @@
-#ifndef COM_NGOS_SHARED_UEFIBASE_SECTIONS_SECTION0_COM_NGOS_SHARED_COMMON_DMI_ENTRY_LIB_DMIBIOSEXTENDEDROMSIZE_H
-#define COM_NGOS_SHARED_UEFIBASE_SECTIONS_SECTION0_COM_NGOS_SHARED_COMMON_DMI_ENTRY_LIB_DMIBIOSEXTENDEDROMSIZE_H
-
-
-
-#include <buildconfig.h>
-#include <com/ngos/shared/common/dmi/lib/entry/lib/dmibiosextendedromsize.h>
-#include <com/ngos/shared/uefibase/testengine.h>
-
-
-
-#if NGOS_BUILD_TEST_MODE == OPTION_YES
-
-
-
-TEST_CASES(section0, com_ngos_shared_common_dmi_lib_entry_lib_dmibiosextendedromsize);
-{
-    TEST_CASE("size()");
-    {
-        DmiBiosExtendedRomSize temp;
-
-
-
-        temp.value = 1;
-        temp.unit  = (u16)DmiBiosExtendedRomSizeUnit::MEGABYTES;
-
-        TEST_ASSERT_EQUALS(temp.size(), MB);
-
-
-
-        temp.value = 64;
-        temp.unit  = (u16)DmiBiosExtendedRomSizeUnit::MEGABYTES;
-
-        TEST_ASSERT_EQUALS(temp.size(), 64 * MB);
-
-
-
-        temp.value = 16383;
-        temp.unit  = (u16)DmiBiosExtendedRomSizeUnit::MEGABYTES;
-
-        TEST_ASSERT_EQUALS(temp.size(), 16383 * MB);
-
-
-
-        temp.value = 1;
-        temp.unit  = (u16)DmiBiosExtendedRomSizeUnit::GIGABYTES;
-
-        TEST_ASSERT_EQUALS(temp.size(), GB);
-
-
-
-        temp.value = 64;
-        temp.unit  = (u16)DmiBiosExtendedRomSizeUnit::GIGABYTES;
-
-        TEST_ASSERT_EQUALS(temp.size(), 64 * GB);
-
-
-
-        temp.value = 16383;
-        temp.unit  = (u16)DmiBiosExtendedRomSizeUnit::GIGABYTES;
-
-        TEST_ASSERT_EQUALS(temp.size(), 16383 * GB);
-    }
-    TEST_CASE_END();
-}
-TEST_CASES_END();
-
-
-
-#endif
-
-
-
-#endif // COM_NGOS_SHARED_UEFIBASE_SECTIONS_SECTION0_COM_NGOS_SHARED_COMMON_DMI_ENTRY_LIB_DMIBIOSEXTENDEDROMSIZE_H
+#ifndef COM_NGOS_SHARED_UEFIBASE_SECTIONS_SECTION0_COM_NGOS_SHARED_COMMON_DMI_ENTRY_LIB_DMIBIOSEXTENDEDROMSIZE_H                                                                                         // Colorize: green
+#define COM_NGOS_SHARED_UEFIBASE_SECTIONS_SECTION0_COM_NGOS_SHARED_COMMON_DMI_ENTRY_LIB_DMIBIOSEXTENDEDROMSIZE_H                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+#include <buildconfig.h>                                                                                                                                                                                 // Colorize: green
+#include <com/ngos/shared/common/dmi/lib/entry/lib/dmibiosextendedromsize.h>                                                                                                                             // Colorize: green
+#include <com/ngos/shared/uefibase/testengine.h>                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+#if NGOS_BUILD_TEST_MODE == OPTION_YES                                                                                                                                                                   // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+TEST_CASES(section0, com_ngos_shared_common_dmi_lib_entry_lib_dmibiosextendedromsize);                                                                                                                   // Colorize: green
+{                                                                                                                                                                                                        // Colorize: green
+    TEST_CASE("size()");                                                                                                                                                                                 // Colorize: green
+    {                                                                                                                                                                                                    // Colorize: green
+        DmiBiosExtendedRomSize temp;                                                                                                                                                                     // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+        temp.value = 1;                                                                                                                                                                                  // Colorize: green
+        temp.unit  = DmiBiosExtendedRomSizeUnit::MEGABYTES;                                                                                                                                              // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+        TEST_ASSERT_EQUALS(temp.size(), MB);                                                                                                                                                             // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+        temp.value = 64;                                                                                                                                                                                 // Colorize: green
+        temp.unit  = DmiBiosExtendedRomSizeUnit::MEGABYTES;                                                                                                                                              // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+        TEST_ASSERT_EQUALS(temp.size(), 64 * MB);                                                                                                                                                        // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+        temp.value = 16383;                                                                                                                                                                              // Colorize: green
+        temp.unit  = DmiBiosExtendedRomSizeUnit::MEGABYTES;                                                                                                                                              // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+        TEST_ASSERT_EQUALS(temp.size(), 16383 * MB);                                                                                                                                                     // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+        temp.value = 1;                                                                                                                                                                                  // Colorize: green
+        temp.unit  = DmiBiosExtendedRomSizeUnit::GIGABYTES;                                                                                                                                              // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+        TEST_ASSERT_EQUALS(temp.size(), GB);                                                                                                                                                             // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+        temp.value = 64;                                                                                                                                                                                 // Colorize: green
+        temp.unit  = DmiBiosExtendedRomSizeUnit::GIGABYTES;                                                                                                                                              // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+        TEST_ASSERT_EQUALS(temp.size(), 64 * GB);                                                                                                                                                        // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+        temp.value = 16383;                                                                                                                                                                              // Colorize: green
+        temp.unit  = DmiBiosExtendedRomSizeUnit::GIGABYTES;                                                                                                                                              // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+        TEST_ASSERT_EQUALS(temp.size(), 16383 * GB);                                                                                                                                                     // Colorize: green
+    }                                                                                                                                                                                                    // Colorize: green
+    TEST_CASE_END();                                                                                                                                                                                     // Colorize: green
+}                                                                                                                                                                                                        // Colorize: green
+TEST_CASES_END();                                                                                                                                                                                        // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+#endif                                                                                                                                                                                                   // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+                                                                                                                                                                                                         // Colorize: green
+#endif // COM_NGOS_SHARED_UEFIBASE_SECTIONS_SECTION0_COM_NGOS_SHARED_COMMON_DMI_ENTRY_LIB_DMIBIOSEXTENDEDROMSIZE_H                                                                                       // Colorize: green
